commentaires/profile

+++ back-end/controllers/article.controller.js
@@ -1,4 +1,5 @@
 const db = require('../config/db');
+const fs = require('fs');
 
 exports.createArticle = (req, res, next) => {
     try {
@@ -11,7 +12,7 @@ exports.createArticle = (req, res, next) => {
         const sql = "INSERT INTO articles (article_userId, article_message, article_img) VALUES (?)";
         db.query(sql, [article], function (error, results) {
             if (!error) {
-                res.status(200).json({ message: 'Création d\'article effectué' });   
+                res.status(200).json({ message: 'Création d\'article effectué' });
             } else {
                 res.status(401).json({ error: 'Erreur creation d\'article' });
                 console.log(error)
@@ -45,14 +46,19 @@ exports.updateArticle = (req, res, next) => {
 }
 
 exports.deleteArticle = (req, res, next) => {
-    const sql = "DELETE FROM articles WHERE article_id=?;";
-    db.query(sql, [req.params.id], function (error, results) {
-        if (!error) {
-            res.status(200).json({ message: 'article supprimé' });
-        } else {
-            res.status(401).json({ error: 'Erreur requete suppression d\'article' });
-        }
-    });
+    try {
+        const sql = "DELETE FROM articles WHERE article_id=? AND article_userId=? ";
+        db.query(sql, [req.params.id, res.locals.userId], function (error, results) {
+            if (!error && results.affectedRows === 1) {
+                res.status(200).json({ message: 'article supprimé' });
+            } else {
+                res.status(401).json({ error: 'Erreur requete suppression d\'article' });
+            }
+        });
+    } catch (error) {
+        res.status(500).json({ error });
+        console.log(error)
+    }
 }
 
 exports.getAllArticle = (req, res, next) => {
